INCOMPLETE Python BMI
