+++ py_cfe/run_cfe_unit_test_mass_balance.py
@@ -2,6 +2,6 @@ import time
 import numpy as np
 import pandas as pd
 import json
-import cfe
-cfe1 = cfe.CFE('./cat_58_config_cfe.json')
+import cfe_orig
+cfe1 = cfe_orig.CFE('./cat_58_config_cfe.json')
 cfe1.run_unit_test()
